requirements.txt
django>=1.11.19
Markdown==2.5.2
MarkupSafe==0.23
Unidecode==0.04.17
argparse==1.2.1
django-bower==5.0.2
django-filter==0.9.2
django-jsonview==1.2.0
djangorestframework==3.9.1
python-slugify==0.0.9
pytz==2014.10
requests>=2.20.0
six>=1.10.0
wsgiref==0.1.2
xlwt==0.7.5
psycopg2==2.7.3
celery==3.1.18
redis==2.10.3
django-import-export==1.2.0
django-authtools==1.4.0
django-tables2==2.0.4
django-celery==3.1.17
XlsxWriter==0.9.6
django-sql-explorer==1.1.1